Simplify class structure
